+++ cmstools-requirements.txt
@@ -2,7 +2,7 @@
 --trusted-host artifactory.algol60.net
 --index-url https://arti.hpc.amslabs.hpecorp.net:443/artifactory/api/pypi/pypi-remote/simple
 --extra-index-url http://artifactory.algol60.net/artifactory/csm-python-modules/simple
--c barebones_image_test-constraints.txt
+-c cmstools-constraints.txt
 boto3
 kubernetes
 packaging
