+++ python-venv/cmstools/lib/api/__init__.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2022, 2024-2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -23,5 +23,7 @@
 #
 
 """
-lib for cmstools test
+API module cmstools tests
 """
+
+from .api import API_GW_SECURE, API_BASE_URL, request, request_and_check_status
