+++ python-venv/cmstools/lib/s3/s3.py
@@ -23,19 +23,18 @@
 #
 
 """
-S3 module for barebones boot test
+S3 module for cmstools test
 """
 
 import base64
 import warnings
 
 import boto3
-import logging
 from urllib3.exceptions import InsecureRequestWarning
 
-from cmstools.lib.common.defs import JsonDict
-from cmstools.lib.common.k8s import get_k8s_secret_data
-from cmstools.lib.common.common_logger import logger
+from cmstools.lib.defs import JsonDict
+from cmstools.lib.k8s import get_k8s_secret_data
+from cmstools.lib.common_logger import logger
 
 from .defs import S3_CREDS_SECRET_FIELDS, S3_CREDS_SECRET_NAME
 from .s3_url import S3Url
