+++ python-venv/cmstools/test/barebones_image_test/__main__.py
@@ -61,10 +61,10 @@ from cmstools.test.barebones_image_test.bos import BosSession, BosTemplate
 from cmstools.test.barebones_image_test.cfs import CfsConfigLayerData, CfsConfig, CfsSession, CfsComponents, CfsComponentUpdateData
 from cmstools.test.barebones_image_test.hsm import ComputeNode, find_compute_node, get_compute_node
 from cmstools.test.barebones_image_test.prodcat import CsmProductCatalogData
-from cmstools.lib.common.defs import ARCH_LIST, TestException as BBException
+from cmstools.lib.defs import ARCH_LIST, CmstoolsException as BBException
 from cmstools.test.barebones_image_test.ims import ImsImage
 from cmstools.test.barebones_image_test.test_resource import TestResource
-from cmstools.lib.common.log import LOG_FILE_PATH
+from cmstools.lib.log import LOG_FILE_PATH
 from .log import logger
