+++ python-venv/cmstools/test/barebones_image_test/bos/bos_session.py
@@ -29,14 +29,14 @@ BosSession and BosSessionStatusFields classes
 from dataclasses import dataclass
 from typing import ClassVar
 
-from cmstools.lib.common.api import request_and_check_status
+from cmstools.lib.api import request_and_check_status
 from cmstools.test.barebones_image_test.hsm import ComputeNode
-from cmstools.lib.common.defs import TestException as BBException
+from cmstools.lib.defs import CmstoolsException as BBException
 from cmstools.test.barebones_image_test.log import logger
 from cmstools.test.barebones_image_test.test_session import SessionStatusFields, TestSession
 
 from .bos_template import BosTemplate
-from cmstools.lib.common.bos.defs import BOS_SESSIONS_URL
+from cmstools.lib.bos.defs import BOS_SESSIONS_URL
 
 
 @dataclass(frozen=True)
