+++ python-venv/cmstools/test/barebones_image_test/cfs/cfs_config.py
@@ -32,11 +32,11 @@ from dataclasses import dataclass
 
 from typing import ClassVar
 
-from cmstools.lib.common.api import request, request_and_check_status
-from cmstools.lib.common.defs import TestException as BBException
+from cmstools.lib.api import request, request_and_check_status
+from cmstools.lib.defs import CmstoolsException as BBException
 from cmstools.test.barebones_image_test.log import logger
 from cmstools.test.barebones_image_test.test_resource import TestResource
-from cmstools.lib.common.cfs.defs import CFS_CONFIGS_URL
+from cmstools.lib.cfs.defs import CFS_CONFIGS_URL
 
 
 @dataclass(frozen=True)
