+++ python-venv/cmstools/test/barebones_image_test/cfs/cfs_session.py
@@ -30,14 +30,14 @@ from dataclasses import dataclass
 
 from typing import ClassVar
 
-from cmstools.lib.common.api import request_and_check_status
-from cmstools.lib.common.defs import TestException as BBException
+from cmstools.lib.api import request_and_check_status
+from cmstools.lib.defs import CmstoolsException as BBException
 from cmstools.test.barebones_image_test.defs import BB_TEST_RESOURCE_NAME
 from cmstools.test.barebones_image_test.ims import ImsImage
 from cmstools.test.barebones_image_test.log import logger
 from cmstools.test.barebones_image_test.test_session import SessionStatusFields, TestSession
 from .cfs_config import CfsConfig
-from cmstools.lib.common.cfs.defs import CFS_SESSIONS_URL
+from cmstools.lib.cfs.defs import CFS_SESSIONS_URL
 
 
 @dataclass(frozen=True)
