+++ python-venv/cmstools/test/barebones_image_test/hsm/compute_node.py
@@ -30,9 +30,9 @@ ComputeNode class and related functions
 from __future__ import annotations
 from dataclasses import dataclass
 
-from cmstools.lib.common.defs import TestException as BBException, JsonDict
+from cmstools.lib.defs import CmstoolsException as BBException, JsonDict
 from cmstools.test.barebones_image_test.log import logger
-from cmstools.lib.common.hsm.defs import HSM_ARCH_STRINGS
+from cmstools.lib.hsm.defs import HSM_ARCH_STRINGS
 
 
 @dataclass(frozen=True)
