+++ python-venv/cmstools/test/barebones_image_test/ims/ims_image.py
@@ -28,12 +28,12 @@ ImsImage class
 
 from typing import ClassVar
 
-from cmstools.lib.common.api import request_and_check_status
-from cmstools.lib.common.defs import TestException as BBException
+from cmstools.lib.api import request_and_check_status
+from cmstools.lib.defs import CmstoolsException as BBException
 from cmstools.test.barebones_image_test.log import logger
-from cmstools.lib.common.s3 import S3Url, get_s3_artifact_etag
+from cmstools.lib.s3 import S3Url, get_s3_artifact_etag
 from cmstools.test.barebones_image_test.test_resource import TestResource
-from cmstools.lib.common.ims.defs import IMS_ARCH_STRINGS, IMS_IMAGES_URL, IMS_V2_IMAGES_URL
+from cmstools.lib.ims.defs import IMS_ARCH_STRINGS, IMS_IMAGES_URL, IMS_V2_IMAGES_URL
 
 # Used to indicate an uninitialized values
 UNSET = object()
