+++ python-venv/cmstools/test/barebones_image_test/prodcat/csm_product_catalog_data.py
@@ -34,10 +34,10 @@ import yaml
 
 import packaging.version
 
-from cmstools.lib.common.defs import TestException as BBException, JsonDict
-from cmstools.lib.common.k8s import get_k8s_configmap_data
+from cmstools.lib.defs import CmstoolsException as BBException, JsonDict
+from cmstools.lib.k8s import get_k8s_configmap_data
 from cmstools.test.barebones_image_test.log import logger
-from cmstools.lib.common.prodcat.defs import PRODCAT_ARCH_STRINGS, PRODUCT_CATALOG_CONFIG_MAP_NAME, \
+from cmstools.lib.prodcat.defs import PRODCAT_ARCH_STRINGS, PRODUCT_CATALOG_CONFIG_MAP_NAME, \
                   PRODUCT_CATALOG_CONFIG_MAP_NS, VCS_URL
